requirements.txt
streamlit
openai
python-pptx
pydantic
pdf2image
Pillow
google-auth
google-auth-oauthlib
python-dotenv
